django-apps/cse535/requirements.txt
Django==2.1.5
djangorestframework==3.10.3
joblib==0.14.0
mysqlclient==1.4.5
numpy==1.17.4
PyMySQL==0.9.3
pytz==2019.3
scikit-learn==0.21.3
scipy==1.3.2
sqlparse==0.3.0
